==> notebooks/02_data_preprocessing.ipynb <==
import pandas as pd
import numpy as np

# Load the synthetic dataset
df = pd.read_csv("../data/raw/bank_transactions.csv", parse_dates=["timestamp"])

print("Shape:", df.shape)
print("\nPreview:")
print(df.head())

print("\nMissing values:")
print(df.isnull().sum())


# Drop duplicate transactions (if any)
df = df.drop_duplicates(subset=["transaction_id"])

# Fill missing categorical values with "Unknown"
categorical_cols = ["merchant_category", "merchant_id", "device_id", "location"]
df[categorical_cols] = df[categorical_cols].fillna("Unknown")

# Fill numeric values with median
df["amount"] = df["amount"].fillna(df["amount"].median())


# Extracting time based features
df["hour_of_day"] = df["timestamp"].dt.hour
df["day_of_week"] = df["timestamp"].dt.dayofweek
df["is_night"] = df["hour_of_day"].apply(lambda x: 1 if x < 6 or x > 22 else 0)
df["is_weekend"] = df["day_of_week"].apply(lambda x: 1 if x >= 5 else 0)


from sklearn.preprocessing import LabelEncoder

label_encoders = {}
categorical_cols = ["merchant_category", "merchant_id", "device_id", "location"]

for col in categorical_cols:
    le = LabelEncoder()
    df[col] = le.fit_transform(df[col])
    label_encoders[col] = le

print("Categorical features encoded successfully.")


# scale data
from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()
df["amount_scaled"] = scaler.fit_transform(df[["amount"]])


df.to_csv("../data/processed/preprocessed_transactions.csv", index=False)
print("✅ Preprocessing complete. Saved as preprocessed_transactions.csv")
