==> notebooks/05_model_trainer.ipynb <==
# Model training - base model
# tracking with mlflow
# Import the specific regression models from scikit-learn
from sklearn.linear_model import LinearRegression, Ridge, Lasso, ElasticNet
import os
from mlflow.models import infer_signature
import pandas as pd
#from urlib.parse import urlparse
import mlflow
#from sklearn.compose import TransformedTargetRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
from dotenv import load_dotenv
from xgboost import XGBRegressor
from lightgbm import LGBMRegressor
from sklearn.neighbors import KNeighborsRegressor
from sklearn.svm import SVR
from sklearn.tree import DecisionTreeRegressor
from sklearn.pipeline import make_pipeline
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import StandardScaler, FunctionTransformer
from sklearn.compose import make_column_selector, make_column_transformer
import numpy as np
import mlflow.sklearn


load_dotenv()

os.environ['MLFLOW_TRACKING_URI']= os.getenv("MLFLOW_TRACKING_URI")
os.environ['MLFLOW_TRACKING_USERNAME']= os.getenv("MLFLOW_TRACKING_USERNAME")
os.environ["MLFLOW_TRACKING_PASSWORD"]= os.getenv("MLFLOW_TRACKING_PASSWORD")

df = pd.read_csv("../data/processed/preprocessed_transactions.csv")
df.head()

df_2 = pd.read_csv("../data/raw/bank_transactions.csv")
df_2.head()

df_2.columns

df.columns

df.dtypes

# Define features
numeric_features = [
    'amount', 'amount_scaled', 'customer_avg_amount',
    'customer_std_amount', 'transaction_frequency',
    'distance_from_home', 'transaction_hour'
]

categorical_features = [
    'device_type', 'location', 'merchant_category', 'channel'
]

# Preprocessing pipeline 

# defining numerical and categorical columns

# preprocessing pipeline the numerical features that is all the features in the dataset
# defining pipeline
num_pipeline = make_pipeline(SimpleImputer(strategy="median"),
                                     StandardScaler())
# pipeline for  the log transformation to handle skew features
log_pipeline = make_pipeline(
    SimpleImputer(strategy="median"),
    FunctionTransformer(np.log, feature_names_out="one-to-one"),
    StandardScaler())



# building the preprocessing pipeline
preprocessing = make_column_transformer(
    (num_pipeline, make_column_selector(dtype_include=np.number)),
    (log_pipeline, make_column_selector(dtype_include=np.number)),
)
